pyproject.toml
[tool.poetry]
name = "book-mirror-plus"
version = "0.1.0"
description = "A local Streamlit + FastAPI app for Goodreads analysis with deep insights"
authors = ["Colby Jackson <colbyjacks1@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}, {include = "ui"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
streamlit = "^1.28.1"
sqlmodel = "^0.0.14"
plotly = "^5.17.0"
faiss-cpu = "^1.7.4"
umap-learn = "^0.5.5"
scikit-learn = "^1.3.2"
pandas = "^2.1.3"
numpy = "^1.25.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
google-generativeai = "^0.8.5"
openai = "^1.3.7"
pydantic = "^2.11.7"
httpx = "^0.25.2"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.1"
pre-commit = "^3.5.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"

[tool.poetry.scripts]
dev = "uvicorn app.api:app --reload --host 0.0.0.0 --port 8000"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.ruff]
line-length = 88
target-version = "py313"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "ARG", "SIM", "TCH", "TID", "Q", "RSE", "RET", "SLF", "SLOT", "PIE", "TRY", "NPY", "AIR", "PERF", "LOG", "FURB", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "AIR", "PERF", "LOG", "FURB", "PTH", "ERA", "PD", "PGH", "PL", "RUF", "S", "BLE", "FBT", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "TID", "TCH", "ARG", "SIM", "UP", "W", "YTT", "S", "BLE", "FBT", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "TID", "TCH", "ARG", "SIM", "UP", "W", "YTT"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto" 